public/requirements.txt
ccxt
pandas
pandas_ta==0.3.14b0  # Utilise la pré-version disponible
psycopg2-binary
flask
numpy<2.0  # Garde numpy<2.0 pour compatibilité